Implemented a proof of concept for base-form reduction
Base-form reduction is the mapping of inflected words onto a
canonical base form. Usually, nominative singular for nouns,
infinitive for verbs and positive predicative for adjectives.
In German, compounds with modifiers based on inflected forms,
e.g. *Ärtzekammern* are a challenge for morphological analyzers
since the correct base form here is *Ärztekammer* and not
*Arztkammer*. The idea of the proof of concept is to build a
second transducer which includes the surface for `deriv_stems`
and `kompos_stems` and the usual deep-surface mapping for
`base_stems`.